Removed ForwardingService class in favor of ControlConstrucService
- all methods touching index with CCs are now in one class
- preparation work before introducing locking
