+++ server/service/individualServices/ControlConstructService.js
@@ -18,6 +18,157 @@ const ElasticsearchPreparation = require('./ElasticsearchPreparation');
 
 class ControlConstructService {
 
+  static async updateForwardingConstruct(controlConstructUuid, newForwardingConstruct) {
+    let esUuid = await ElasticsearchPreparation.getCorrectEsUuid(false);
+    let client = await elasticsearchService.getClient(false, esUuid);
+    let indexAlias = await getIndexAliasAsync(esUuid);
+    let response = await client.updateByQuery({
+      index: indexAlias,
+      refresh: true,
+      body: {
+        script: {
+          source: `ctx._source['forwarding-domain'][0]['forwarding-construct'].removeIf(fc -> fc['uuid'] == params['uuid']);
+                    ctx._source['forwarding-domain'][0]['forwarding-construct'].add(params['newForwardingConstruct'])`,
+          params: {
+            "newForwardingConstruct": newForwardingConstruct,
+            "uuid": newForwardingConstruct[onfAttributes.GLOBAL_CLASS.UUID]
+          }
+        },
+        query: {
+          term: {
+            "uuid": controlConstructUuid
+          }
+        }
+      },
+    });
+    if (response.body.updated === 1) {
+      return { "took": response.body.took };
+    } else {
+      if (response.body.total === 0) {
+        throw new createHttpError.BadRequest(`CC with uuid ${controlConstructUuid} does not exist.`)
+      }
+      throw new Error("Forwarding Construct was not updated")
+    }
+  }
+
+  static async updateFCPort(controlConstructUuid, fcUuid, newFCPort) {
+    let esUuid = await ElasticsearchPreparation.getCorrectEsUuid(false);
+    let client = await elasticsearchService.getClient(false, esUuid);
+    let indexAlias = await getIndexAliasAsync(esUuid);
+    let response = await client.updateByQuery({
+      index: indexAlias,
+      refresh: true,
+      body: {
+        script: {
+          source: `def fwDomain = ctx._source['forwarding-domain'];
+                for (domain in fwDomain) {
+                    def fcs = domain['forwarding-construct'];
+                    for (fc in fcs) {
+                    if (fc['uuid'] == params['fc-uuid']) {
+                        def ports = fc['fc-port'];
+                        ports.removeIf(port -> port['local-id'] == params['local-id']);
+                        ports.add(params['new-fc-port']);
+                    }
+                    }
+                }`,
+          params: {
+            "new-fc-port": newFCPort,
+            "local-id": newFCPort[onfAttributes.LOCAL_CLASS.LOCAL_ID],
+            "fc-uuid": fcUuid
+          }
+        },
+        query: {
+          term: {
+            "uuid": controlConstructUuid
+          }
+        }
+      }
+    });
+    if (response.body.updated === 1) {
+      return { "took": response.body.took };
+    } else {
+      if (response.body.total === 0) {
+        throw new createHttpError.BadRequest(`CC with uuid ${controlConstructUuid} does not exist.`)
+      }
+      throw new Error("FCPort was not updated")
+    }
+  }
+
+  /**
+   * @description Removes fcPort from forwarding-construct
+   * @param {String} fcPortLocalId Local ID of fcPort that should be deleted
+   * @param {String} forwardingConstructUuid UUID of forwarding-construct containing fc-port to be deleted
+   * @param {String} controlConstructUuid UUID of affected control construct
+   * @returns {Promise<Object>} { took }
+   */
+  static async deleteFcPort(fcPortLocalId, forwardingConstructUuid, controlConstructUuid) {
+    let esUuid = await ElasticsearchPreparation.getCorrectEsUuid(false);
+    let client = await elasticsearchService.getClient(false, esUuid);
+    let indexAlias = await getIndexAliasAsync(esUuid);
+    let response = await client.updateByQuery(
+      {
+        index: indexAlias,
+        refresh: true,
+        body: {
+          "script": {
+            "source": `def fwDomain = ctx._source['forwarding-domain'];
+                for (domain in fwDomain) {
+                    def fcs = domain['forwarding-construct'];
+                    for (fc in fcs) {
+                    if (fc['uuid'] == params['fc-uuid']) {
+                        def ports = fc['fc-port'];
+                        ports.removeIf(port -> port['local-id'] == params['local-id'])
+                    }
+                    }
+                }
+                `,
+            "params": {
+              "local-id": fcPortLocalId,
+              "fc-uuid": forwardingConstructUuid
+            }
+          },
+          "query": {
+            "term": {
+              "uuid": controlConstructUuid
+            }
+          }
+        }
+      }
+    );
+    return { "took": response.body.took };
+  }
+
+  /**
+   * @description Finds and deletes fc-ports where INPUT ports match given ltpUUID.
+   * @param {Object} controlConstruct
+   * @param {String} ltpUuid
+   * @returns {Promise<Object>} { took }
+   */
+  static async deleteDependentFcPorts(controlConstruct, ltpUuid) {
+    let forwardingDomainList = controlConstruct[onfAttributes.CONTROL_CONSTRUCT.FORWARDING_DOMAIN];
+    let took = 0;
+    for (let forwardingDomain of forwardingDomainList) {
+      let fcList = forwardingDomain[onfAttributes.FORWARDING_DOMAIN.FORWARDING_CONSTRUCT];
+      for (let fc of fcList) {
+        let fcUuid = fc[onfAttributes.GLOBAL_CLASS.UUID];
+        let fcPortList = fc[onfAttributes.FORWARDING_CONSTRUCT.FC_PORT];
+        for (let fcPort of fcPortList) {
+          let fcPortLocalId = fcPort[onfAttributes.LOCAL_CLASS.LOCAL_ID];
+          let fcPortlogicalTerminationPoint = fcPort[onfAttributes.CONTROL_CONSTRUCT.LOGICAL_TERMINATION_POINT];
+          if (fcPortlogicalTerminationPoint === ltpUuid) {
+            let deleteResponse = this.deleteFcPort(
+              fcPortLocalId,
+              fcUuid,
+              controlConstruct[onfAttributes.GLOBAL_CLASS.UUID]
+            );
+            took += deleteResponse.took;
+          }
+        }
+      }
+    }
+    return { "took": took };
+  }
+
   /**
    * @description This function returns the control-construct from Elasticsearch.
    * @param {String} controlConstructUuid
@@ -68,10 +219,10 @@ class ControlConstructService {
     });
     if (res.body.hits === undefined) {
       console.log(`Could not find existing control-construct with LTP UUID ${ltpUuid}`);
-      return { "took" : res.body.took };
+      return { "took": res.body.took };
     }
     let controlConstruct = createResultArray(res);
-    return { "controlConstruct" : controlConstruct[0], "took" : res.body.took };
+    return { "controlConstruct": controlConstruct[0], "took": res.body.took };
   }
 
   /**
@@ -95,9 +246,9 @@ class ControlConstructService {
       }
     });
     if (res.body.hits === undefined) {
-      return { "took" : res.body.took };
+      return { "took": res.body.took };
     }
-    return { "id": res.body.hits.hits[0]._id, "took" : res.body.took }
+    return { "id": res.body.hits.hits[0]._id, "took": res.body.took }
   }
 
   /**
@@ -190,7 +341,7 @@ class ControlConstructService {
     ltps.forEach(ltp => {
       let serverLtps = ltp[onfAttributes.LOGICAL_TERMINATION_POINT.SERVER_LTP];
       if (serverLtps.length !== 0) {
-        let serverLtpIndex = serverLtps.findIndex(serverLtp =>serverLtp === ltpToBeRemovedUuid);
+        let serverLtpIndex = serverLtps.findIndex(serverLtp => serverLtp === ltpToBeRemovedUuid);
         if (serverLtpIndex !== -1) {
           serverLtps.splice(serverLtpIndex, 1);
         }
@@ -210,25 +361,25 @@ class ControlConstructService {
     let client = await elasticsearchService.getClient(false, esUuid);
     let indexAlias = await getIndexAliasAsync(esUuid);
     let res = await client.search({
-        index: indexAlias,
-        filter_path: 'took,**.http-server-interface-capability.application-name,' +
+      index: indexAlias,
+      filter_path: 'took,**.http-server-interface-capability.application-name,' +
         '**.http-server-interface-capability.release-number',
-        body: {
-            "query": {
-            "term": {"logical-termination-point.uuid": ltpUuid}
-            }
+      body: {
+        "query": {
+          "term": { "logical-termination-point.uuid": ltpUuid }
         }
+      }
     });
     if (res.body.hits.length === 0) {
-        throw new Error('Http server capability not found!');
+      throw new Error('Http server capability not found!');
     }
     let filteredLtps = res.body.hits.hits[0]._source['logical-termination-point'];
     for (let ltp of filteredLtps) {
-        let layerProtocol = ltp[onfAttributes.LOGICAL_TERMINATION_POINT.LAYER_PROTOCOL][0];
-        if (onfAttributes.LAYER_PROTOCOL.HTTP_SERVER_INTERFACE_PAC in layerProtocol) {
-          let httpServerCapability = layerProtocol[onfAttributes.LAYER_PROTOCOL.HTTP_SERVER_INTERFACE_PAC][onfAttributes.HTTP_SERVER.CAPABILITY];
-            return { "httpServerCapability": httpServerCapability, "took": res.body.took };
-        }
+      let layerProtocol = ltp[onfAttributes.LOGICAL_TERMINATION_POINT.LAYER_PROTOCOL][0];
+      if (onfAttributes.LAYER_PROTOCOL.HTTP_SERVER_INTERFACE_PAC in layerProtocol) {
+        let httpServerCapability = layerProtocol[onfAttributes.LAYER_PROTOCOL.HTTP_SERVER_INTERFACE_PAC][onfAttributes.HTTP_SERVER.CAPABILITY];
+        return { "httpServerCapability": httpServerCapability, "took": res.body.took };
+      }
     };
     return { "took": res.body.took };
   }
